# GestaoHotelariaAPI/GestaoHotelariaAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GestaoHotelariaAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Utilizadores": {
      "get": {
        "tags": [
          "Utilizadores"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UtilizadorListServiceResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Utilizadores"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Utilizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UtilizadorListServiceResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Utilizadores"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Utilizador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UtilizadorListServiceResponse"
            }
          }
        }
      }
    },
    "/api/Utilizadores/{id}": {
      "get": {
        "tags": [
          "Utilizadores"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UtilizadorServiceResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Utilizadores"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UtilizadorListServiceResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Utilizador": {
      "type": "object",
      "properties": {
        "utilizadorID": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "perfil": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UtilizadorListServiceResponse": {
      "type": "object",
      "properties": {
        "dados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Utilizador"
          }
        },
        "mensagem": {
          "type": "string"
        },
        "sucesso": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "UtilizadorServiceResponse": {
      "type": "object",
      "properties": {
        "dados": {
          "$ref": "#/definitions/Utilizador"
        },
        "mensagem": {
          "type": "string"
        },
        "sucesso": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT no formato 'Bearer {token}'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}